refactor: H3 dashboard decomposition + M1/M2 constants (#176)
* refactor: fix formatCurrency locale and add makeCurrencyFormatter helper


* refactor: add shared dateOptions constants (yearOptions, monthOptions)


* refactor: extract TrendBadge to its own component file


* refactor: extract TypeSummaryCards component


* refactor: extract CategoryBreakdownSection component


* refactor: extract PeriodTab; index.tsx is now a 55-line shell


* refactor: extract TrendChart component


* refactor: extract SpendingTrendlineChart component


* refactor: extract TagBar; ChartsTab.tsx is now a 60-line shell


* refactor: replace useList+useMemo with useSelect for tags

Replace the useList + useMemo pattern for tags with useSelect hook from
@refinedev/core in both create and edit transaction pages. This simplifies
the code by letting useSelect handle the options mapping internally.


* fix: widen month selectors in ChartsTab for full month names


* docs: update H1/H2/H3 spec status; add H3 implementation plan


* fix: remove unused currentYear import in ChartsTab


* docs: mark H3/M1/M2 complete in plan and spec

- All 58 task checkboxes in plan file checked
- Plan header updated with completion status and PR #176 reference
- Spec: status line updated (H3/M1/M2 ✅)
- Spec: H3/M1/M2 headings each get COMPLETE + status block citing PR #176
- Spec: implementation order strikethroughs for M2/H3/M1
- Spec: appendix updated with post-H3 line counts (index.tsx 579→54, ChartsTab.tsx 587→~140)


* fix: address PR #176 review feedback

- dateOptions.ts: use startOf('year').month(i) to prevent month
  overflow when today's day-of-month > days in target month
- transactions/create.tsx: rename core useSelect import to useCoreSelect
  to disambiguate from @refinedev/antd's useSelect (useAntSelect)
- transactions/edit.tsx: same rename


---------
