+++ apps/web-next/src/constants/dateOptions.ts
@@ -1,0 +1,13 @@
+import dayjs from "dayjs";
+
+export const currentYear = dayjs().year();
+
+export const yearOptions = Array.from({ length: 6 }, (_, i) => ({
+  label: String(currentYear - i),
+  value: currentYear - i,
+}));
+
+export const monthOptions = Array.from({ length: 12 }, (_, i) => ({
+  label: dayjs().startOf("year").month(i).format("MMMM"),
+  value: i,
+}));
