+++ apps/web-next/src/pages/dashboard/components/TagBar.tsx
@@ -1,0 +1,64 @@
+import { Card, Typography } from "antd";
+import {
+  BarChart,
+  Bar,
+  XAxis,
+  YAxis,
+  CartesianGrid,
+  Tooltip,
+  ResponsiveContainer,
+} from "recharts";
+import { makeCurrencyFormatter } from "../../../utility/currency";
+
+const { Text } = Typography;
+
+export const TagBar = ({
+  title,
+  data,
+  color,
+  currency,
+}: {
+  title: string;
+  data: { tag: string; total: number }[];
+  color: string;
+  currency: string;
+}) => {
+  const fmt = makeCurrencyFormatter(currency);
+  return (
+    <Card title={title}>
+      {data.length === 0 ? (
+        <Text type="secondary">No tagged transactions in this period</Text>
+      ) : (
+        <ResponsiveContainer
+          width="100%"
+          height={Math.max(160, data.length * 34)}
+        >
+          <BarChart
+            layout="vertical"
+            data={data}
+            margin={{ top: 4, right: 32, left: 16, bottom: 4 }}
+          >
+            <CartesianGrid
+              strokeDasharray="3 3"
+              stroke="#f0f0f0"
+              horizontal={false}
+            />
+            <XAxis
+              type="number"
+              tickFormatter={(v) => fmt(v)}
+              tick={{ fontSize: 11 }}
+            />
+            <YAxis
+              type="category"
+              dataKey="tag"
+              tick={{ fontSize: 12 }}
+              width={96}
+            />
+            <Tooltip formatter={fmt} />
+            <Bar dataKey="total" fill={color} radius={[0, 3, 3, 0]} />
+          </BarChart>
+        </ResponsiveContainer>
+      )}
+    </Card>
+  );
+};
