+++ apps/web-next/src/utility/currency.ts
@@ -3,9 +3,9 @@ export const formatCurrency = (
   currency = "USD"
 ): string => {
   const num = typeof amount === "string" ? parseFloat(amount) : amount;
-  return new Intl.NumberFormat("en-US", {
+  return new Intl.NumberFormat(undefined, {
     style: "currency",
-    currency: currency,
+    currency,
   }).format(num);
 };
 
@@ -13,3 +13,9 @@ export const formatAmount = (amount: number | string): string => {
   const num = typeof amount === "string" ? parseFloat(amount) : amount;
   return num.toFixed(2);
 };
+
+/** Returns a formatter function suitable for Recharts `formatter` prop. */
+export const makeCurrencyFormatter =
+  (currency: string) =>
+  (value: unknown): string =>
+    formatCurrency(typeof value === "number" ? value : 0, currency);
